+++ batch/hotcrapi.php
@@ -1,5 +1,5 @@
 <?php
-// hotcli.php -- HotCRP script for interacting with site APIs
+// Hotcrapi.php -- Hotcrapi command-line tool interacts with HotCRP APIs
 // Copyright (c) 2006-2025 Eddie Kohler; see LICENSE.
 
 if (realpath($_SERVER["PHP_SELF"]) === __FILE__) {
@@ -8,17 +8,17 @@ if (realpath($_SERVER["PHP_SELF"]) === __FILE__) {
 
 interface CLIBatchCommand {
     /** @return int */
-    function run(HotCLI_Batch $clib);
+    function run(Hotcrapi_Batch $clib);
 }
 
-class HotCLI_Batch_Site {
+class Hotcrapi_Batch_Site {
     /** @var string */
     public $site;
     /** @var string */
     public $apitoken;
 }
 
-class HotCLI_File {
+class Hotcrapi_File {
     /** @var resource */
     public $stream;
     /** @var string */
@@ -29,12 +29,12 @@ class HotCLI_File {
     public $size;
 
     /** @param string $fn
-     * @return HotCLI_File */
+     * @return Hotcrapi_File */
     static function make($fn, $mode = "rb") {
         if ($fn === "") {
             throw new CommandLineException("Empty filename");
         }
-        $cf = new HotCLI_File;
+        $cf = new Hotcrapi_File;
         if ($fn === "-") {
             $cf->stream = STDIN;
             $cf->filename = null;
@@ -50,9 +50,21 @@ class HotCLI_File {
         }
         return $cf;
     }
+
+    /** @param string $data
+     * @return Hotcrapi_File */
+    static function make_data($data) {
+        $cf = new Hotcrapi_File;
+        $cf->stream = fopen("php://temp", "w+b");
+        fwrite($cf->stream, $data);
+        rewind($cf->stream);
+        $cf->input_filename = "<data>";
+        $cf->size = strlen($data);
+        return $cf;
+    }
 }
 
-class HotCLI_Batch extends MessageSet {
+class Hotcrapi_Batch extends MessageSet {
     /** @var string
      * @readonly */
     public $site;
@@ -85,9 +97,9 @@ class HotCLI_Batch extends MessageSet {
     /** @var ?string */
     private $output;
 
-    /** @var array<string,HotCLI_Batch_Site> */
+    /** @var array<string,Hotcrapi_Batch_Site> */
     private $siteconfig = [];
-    /** @var ?HotCLI_Batch_Site */
+    /** @var ?Hotcrapi_Batch_Site */
     private $default_siteconfig;
 
     function __construct() {
@@ -161,7 +173,7 @@ class HotCLI_Batch extends MessageSet {
             return $this;
         }
         if (!isset($f)) {
-            $f = getenv("HOME") . "/.hotcliconfig";
+            $f = getenv("HOME") . "/.hotcrapi.conf";
             if (!file_exists($f)) {
                 return $this;
             }
@@ -183,7 +195,7 @@ class HotCLI_Batch extends MessageSet {
             ++$line;
             if (preg_match('/\A\s*+\[\s*+site\s*+(\w+|\".*?\"|(?=\]))\s*+\]\s*+\z/', $l, $m)) {
                 $sn = self::unquote($m[1]);
-                $this->siteconfig[$sn] = $this->siteconfig[$sn] ?? new HotCLI_Batch_Site;
+                $this->siteconfig[$sn] = $this->siteconfig[$sn] ?? new Hotcrapi_Batch_Site;
             } else if (preg_match('/\A\s*+\[/', $l)) {
                 $sn = null;
             } else if ($sn === null) {
@@ -204,18 +216,27 @@ class HotCLI_Batch extends MessageSet {
                 $this->default_siteconfig = $this->siteconfig[$sn];
             }
         }
+
+        if (($dsc = $this->default_siteconfig)) {
+            if ($this->site !== null && $dsc->site !== null) {
+                $this->site = $dsc->site;
+            }
+            if ($this->apitoken !== null && $dsc->apitoken !== null) {
+                $this->set_apitoken($dsc->apitoken);
+            }
+        }
     }
 
-    /** @return ?HotCLI_Batch_Site */
+    /** @return ?Hotcrapi_Batch_Site */
     function default_site() {
         return $this->default_siteconfig;
     }
 
-    /** @param string|HotCLI_Batch_Site $s
+    /** @param string|Hotcrapi_Batch_Site $s
      * @return $this
      * @suppress PhanAccessReadOnlyProperty */
     function set_site($s) {
-        if ($s instanceof HotCLI_Batch_Site) {
+        if ($s instanceof Hotcrapi_Batch_Site) {
             if ($s->site !== null) {
                 $this->site = $s->site;
             }
@@ -261,7 +282,7 @@ class HotCLI_Batch extends MessageSet {
         return $this->apitoken !== null;
     }
 
-    /** @param ?callable(HotCLI_Batch):bool $skip_function
+    /** @param ?callable(Hotcrapi_Batch):bool $skip_function
      * @return bool */
     function exec_api($skip_function = null) {
         if ($this->verbose) {
@@ -318,10 +339,10 @@ class HotCLI_Batch extends MessageSet {
     /** @return int */
     function run() {
         if (!$this->has_site()) {
-            throw new CommandLineException("`-s SITE` required");
+            throw new CommandLineException("`-S SITE` required");
         }
         if (!$this->has_apitoken()) {
-            throw new CommandLineException("`-t APITOKEN` required");
+            throw new CommandLineException("`-T APITOKEN` required");
         }
         $status = $this->command->run($this);
         if (!$this->quiet) {
@@ -334,60 +355,45 @@ class HotCLI_Batch extends MessageSet {
     }
 
     /** @param list<string> $argv
-     * @return HotCLI_Batch */
+     * @return Hotcrapi_Batch */
     static function make_args($argv) {
         $getopt = (new Getopt)->long(
             "help::,h:: !",
+            "S:,siteurl:,url: =SITE Site URL",
+            "T:,token: =APITOKEN API token",
+            "F:,config: =FILE Set configuration file [~/.hotcrapi.conf]",
             "verbose,V Be verbose",
-            "F:,config: =FILE Set configuration file",
-            "s:,siteurl:,url:,u: =SITE Site URL",
-            "t:,token: =APITOKEN API token",
-            "filename:,f: =FILENAME !upload Filename for uploaded file",
-            "no-filename !",
-            "mimetype:,m: =MIMETYPE !upload Type for uploaded file",
-            "p:,paper: =PID !paper Submission ID",
-            "q:,query: =SEARCH !paper Submission search",
-            "e,edit !paper Change submissions",
-            "delete !paper Delete submission",
-            "dry-run,d !paper Don’t actually edit",
-            "disable-users !paper Disable newly created users",
-            "add-topics !paper Add all referenced topics to conference",
-            "reason: !paper Reason for update (included in notifications)",
-            "no-notify Don’t notify users",
-            "no-notify-authors Don’t notify authors",
-            "chunk: =CHUNKSIZE Maximum upload chunk size [8M]",
             "quiet Do not print error messages",
-        )->subcommand(true,
-            "upload Upload file to HotCRP and return token",
-            "paper Retrieve or change HotCRP submissions"
+            "chunk: =CHUNKSIZE Maximum upload chunk size [8M]"
         )->description("Interact with HotCRP site using APIs
-Usage: php batch/hotcli.php -u SITEURL -t APITOKEN SUBCOMMAND ARGS...")
-         ->helpopt("help");
+Usage: php batch/hotcrapi.php -S SITEURL -T APITOKEN SUBCOMMAND ARGS...")
+         ->helpopt("help")
+         ->subcommand(true);
+        Paper_CLIBatch::register_options($getopt);
+        Upload_CLIBatch::register_options($getopt);
         $arg = $getopt->parse($argv);
 
-        $hcli = new HotCLI_Batch;
+        $hcli = new Hotcrapi_Batch;
         $hcli->load_config_file($arg["F"] ?? null);
 
-        if (isset($arg["s"])) {
-            $hcli->set_site($arg["s"]);
-        } else if (isset($_ENV["HOTCLI_SITE"])) {
-            $hcli->set_site($_ENV["HOTCLI_SITE"]);
-        } else if (($s = $hcli->default_site())) {
-            $hcli->set_site($s);
+        if (isset($arg["S"])) {
+            $hcli->set_site($arg["S"]);
+        } else if (isset($_ENV["HOTCRAPI_SITE"])) {
+            $hcli->set_site($_ENV["HOTCRAPI_SITE"]);
         }
 
-        if (isset($arg["t"])) {
-            if (str_starts_with($arg["t"], "<")) {
-                $t = @file_get_contents(substr($arg["t"], 1));
+        if (isset($arg["T"])) {
+            if (str_starts_with($arg["T"], "<")) {
+                $t = @file_get_contents(substr($arg["T"], 1));
                 if ($t === false) {
-                    throw CommandLineException::make_file_error(substr($arg["t"], 1));
+                    throw CommandLineException::make_file_error(substr($arg["T"], 1));
                 }
             } else {
-                $t = $arg["t"];
+                $t = $arg["T"];
             }
             $hcli->set_apitoken($t);
-        } else if (isset($_ENV["HOTCLI_APITOKEN"])) {
-            $hcli->set_apitoken($_ENV["HOTCLI_APITOKEN"]);
+        } else if (isset($_ENV["HOTCRAPI_TOKEN"])) {
+            $hcli->set_apitoken($_ENV["HOTCRAPI_TOKEN"]);
         }
 
         if (isset($arg["quiet"])) {
@@ -426,5 +432,5 @@ Usage: php batch/hotcli.php -u SITEURL -t APITOKEN SUBCOMMAND ARGS...")
 }
 
 if (realpath($_SERVER["PHP_SELF"]) === __FILE__) {
-    exit(HotCLI_Batch::make_args($argv)->run());
+    exit(Hotcrapi_Batch::make_args($argv)->run());
 }